+++ packages/core/src/react/index.tsx
@@ -244,4 +244,4 @@ export const Overview: React.FC<OverviewProps> = ({ observed, className, style }
   );
 };
 
-export const version: string = '1.2.0';
+export const version: string = '1.3.0';