+++ packages/core/test/react.test.tsx
@@ -49,7 +49,7 @@ function renderStateful<P>(
 
 describe('graphojs/react', () => {
   it('exposes a version', () => {
-    expect(version).toBe('1.2.0');
+    expect(version).toBe('1.3.0');
   });
 
   it('renders a Diagram component and initializes the diagram', async () => {